Change m_y_delta to float

+++ include/AutomationEditor.h
@@ -257,7 +257,7 @@ private:
 	tick_t m_drawLastTick;
 
 	int m_ppb;
-	int m_y_delta;
+	float m_y_delta;
 	bool m_y_auto;
 
 	// Time position (key) of automation node whose outValue is being dragged
